more code cleanup

+++ Configuration/TCA/tx_t3events_domain_model_company.php
@@ -100,7 +100,7 @@ return [
                 'checkbox' => 0,
                 'default' => 0,
                 'range' => [
-                    'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y'))
+                    'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y'))
                 ],
                 'behaviour' => [
                     'allowLanguageSynchronization' => true
@@ -118,7 +118,7 @@ return [
                 'checkbox' => 0,
                 'default' => 0,
                 'range' => [
-                    'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y'))
+                    'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y'))
                 ],
                 'behaviour' => [
                     'allowLanguageSynchronization' => true
