+++ Configuration/TCA/tx_t3events_domain_model_ticketclass.php
@@ -111,7 +111,7 @@ return [
                 'checkbox' => 0,
                 'default' => 0,
                 'range' => [
-                    'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y'))
+                    'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y'))
                 ],
                 'behaviour' => [
                     'allowLanguageSynchronization' => true
@@ -128,7 +128,7 @@ return [
                 'checkbox' => 0,
                 'default' => 0,
                 'range' => [
-                    'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y'))
+                    'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y'))
                 ],
                 'behaviour' => [
                     'allowLanguageSynchronization' => true
